// src/types/EachJobType.ts
 type EachJobType = {
    _id?: string;
    employerId?: string;
    company?: string;
    jobTitle?: string;
    jobType?: string;
    location?: string;
    minSalary?: number;
    maxSalary?: number;
    responsibilities?: string;
    requirements?: string;
    skills?: string;
    minExp?: number;
    maxExp?: number;
    openings?: number;
    __v?: number;
    postedOn?: string;
 };

 export default EachJobType;